Add stream executor c api extension
